Rename log to userLog
